updated the kernel function to use pass by reference instead of by value

+++ Sqnn/src/KdeUsingMrpt.h
@@ -16,7 +16,7 @@ public:
 	//TODO implementation
 	//Should only handle allocation. Make method for isValid after.
 	//TODO: include and save lambda for distance
-	KdeUsingMrpt(Eigen::MatrixXf &data, int n, int d, int k, int samples, int trees, float sigma,
+	KdeUsingMrpt(Eigen::MatrixXf &data, int n, int d, int k, int samples, int trees, double sigma,
 							 kernel::kernelLambda<float> kernel)
 			: data(data), n(n), d(d), samples(samples), KNN(k), sigma(sigma), mrpt(data), kernel(std::move(kernel)) {
